output_parsers/react_single_Agent: Fixed edge case wrt parsing output from LLM for pandas_agent
The LLM does not always get "sensible" questions, and in those cases,
there is no tools in the Action (Action: None) and no final answer.
This was not handled correctly by the parsers,
leading to OutputParseException.

To avoid unnecessary Exceptions, such answers are now considered
valid and the "though", where the LLM reasons that it is not
capable of answering becomes the final answer.